+++ common/src/main/java/com/lightstep/tracer/shared/Options.java
@@ -214,7 +214,7 @@ public final class Options {
         private Map<Format<?>, Propagator> propagators = new HashMap<>();
         private boolean disableMetaEventLogging = false;
         private String metricsUrl = LightStepConstants.Metrics.DEFAULT_URL;
-        private boolean disableMetricsReporting = false;
+        private boolean disableMetricsReporting = getEnvMetricsDisabled();
         private CollectorClient collectorClient;
         private String grpcCollectorTarget;
         private boolean grpcRoundRobin = false;
@@ -247,6 +247,14 @@ public final class Options {
             this.grpcCollectorTarget = options.grpcCollectorTarget;
             this.grpcRoundRobin = options.grpcRoundRobin;
             this.okhttpDns = options.okhttpDns;
+        }
+
+        private static boolean getEnvMetricsDisabled() {
+            String metricEnabled = System.getenv(LightStepConstants.Metrics.LS_METRICS_ENABLED);
+            if (metricEnabled != null) {
+                return "false".equals(metricEnabled);
+            }
+            return LightStepConstants.Metrics.DEFAULT_DISABLE_METRICS;
         }
 
         /**