+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 <a name="Pending Release"></a>
 ## [Pending Release](https://github.com/lightstep/lightstep-tracer-java-common/compare/0.30.0...master)
+* Allow users to disable metrics via environment variable: `LS_METRICS_ENABLED=false`
 
 <a name="0.30.0"></a>
 ## [0.30.0](https://github.com/lightstep/lightstep-tracer-java-common/compare/0.21.0...0.30.0)